+++ lib/rspec/mocks.rb
@@ -35,6 +35,25 @@ module RSpec
         warn(message)
       end
 
+      # @api private
+      KERNEL_METHOD_METHOD = ::Kernel.instance_method(:method)
+
+      # @api private
+      # Used internally to get a method handle for a particular object
+      # and method name.
+      #
+      # Includes handling for a few special cases:
+      #
+      #   - Objects that redefine #method (e.g. an HTTPRequest struct)
+      #   - BasicObject subclasses that mixin a Kernel dup (e.g. SimpleDelegator)
+      def method_handle_for(object, method_name)
+        if ::Kernel === object
+          KERNEL_METHOD_METHOD.bind(object).call(method_name)
+        else
+          object.method(method_name)
+        end
+      end
+
     private
 
       def add_extensions