+++ lib/rspec/mocks/any_instance/recorder.rb
@@ -180,7 +180,7 @@ module RSpec
           backup_method!(method_name)
           @klass.class_eval(<<-EOM, __FILE__, __LINE__ + 1)
             def #{method_name}(*args, &blk)
-              klass = ::Kernel.instance_method(:method).bind(self).call(:#{method_name}).owner
+              klass = ::RSpec::Mocks.method_handle_for(self, :#{method_name}).owner
               klass.__recorder.playback!(self, :#{method_name})
               self.__send__(:#{method_name}, *args, &blk)
             end
@@ -192,7 +192,7 @@ module RSpec
           @klass.class_eval(<<-EOM, __FILE__, __LINE__ + 1)
             def #{method_name}(*args, &blk)
               method_name = :#{method_name}
-              klass = ::Kernel.instance_method(:method).bind(self).call(:#{method_name}).owner
+              klass = ::RSpec::Mocks.method_handle_for(self, :#{method_name}).owner
               invoked_instance = klass.__recorder.instance_that_received(method_name)
               raise RSpec::Mocks::MockExpectationError, "The message '#{method_name}' was received by \#{self.inspect} but has already been received by \#{invoked_instance}"
             end